requirements.txt
appnope==0.1.2
backcall==0.2.0
click==8.0.3
debugpy==1.5.1
decorator==5.1.0
entrypoints==0.3
geomet==0.2.1.post1
ipykernel==6.6.0
ipython==7.30.1
jedi==0.18.1
jupyter-client==7.1.0
jupyter-core==4.9.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
numpy==1.21.4
pandas==1.3.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.24
psycopg2-binary==2.9.2
ptyprocess==0.7.0
Pygments==2.10.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pyzmq==22.3.0
six==1.16.0
tornado==6.1
traitlets==5.1.1
wcwidth==0.2.5
